+++ Code/Maintenance System/nbproject/private/private.properties
@@ -1,14 +1,14 @@
 <<<<<<< HEAD
 deploy.ant.properties.file=C:\\Users\\ADMIN\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
 =======
-deploy.ant.properties.file=C:\\Users\\sonNH\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
+deploy.ant.properties.file=C:\\Users\\ADMIN\\AppData\\Roaming\\NetBeans\\17\\tomcat101.properties
 file.reference.jakarta.activation-2.0.1.jar=D:\\SWP_ProductsMaintainManagement\\ProductsMaintainManagement (main)\\lib\\jakarta.activation-2.0.1.jar
 file.reference.jakarta.mail-2.0.1.jar=D:\\SWP_ProductsMaintainManagement\\ProductsMaintainManagement (main)\\lib\\jakarta.mail-2.0.1.jar
 >>>>>>> e569b45d1b0c88856a5c908f9281679bc04a8039
-j2ee.server.home=C:/Program Files/Apache Software Foundation/Tomcat 10.1
-j2ee.server.instance=tomcat101:home=C:\\Program Files\\Apache Software Foundation\\Tomcat 10.1
+j2ee.server.home=D:/app_NetBeans/TomCat
+j2ee.server.instance=tomcat101:home=D:\\app_NetBeans\\TomCat
 
 javac.debug=true
 javadoc.preview=true
 selected.browser=default
-user.properties.file=C:\\Users\\sonNH\\AppData\\Roaming\\NetBeans\\17\\build.properties
+user.properties.file=C:\\Users\\ADMIN\\AppData\\Roaming\\NetBeans\\17\\build.properties
