+++ Code/Maintenance System/src/java/Controller/Feedback/CreateFeedback.java
@@ -3,12 +3,9 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
 
-package Controller;
+package Controller.Feedback;
 
 import DAO.FeedbackDAO;
-import DAO.FeedbackLogDAO;
-import Model.Feedback;
-import Model.FeedbackLog;
 import java.io.IOException;
 import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
@@ -16,14 +13,13 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import java.util.ArrayList;
 
 /**
  *
  * @author Tra Pham
  */
-@WebServlet(name="ViewListFeedback", urlPatterns={"/ViewListFeedback"})
-public class ViewListFeedback extends HttpServlet {
+@WebServlet(name="CreateFeedback", urlPatterns={"/CreateFeedback"})
+public class CreateFeedback extends HttpServlet {
    
     /** 
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
@@ -40,10 +36,10 @@ public class ViewListFeedback extends HttpServlet {
             out.println("<!DOCTYPE html>");
             out.println("<html>");
             out.println("<head>");
-            out.println("<title>Servlet ViewListFeedback</title>");  
+            out.println("<title>Servlet CreateFeedback</title>");  
             out.println("</head>");
             out.println("<body>");
-            out.println("<h1>Servlet ViewListFeedback at " + request.getContextPath () + "</h1>");
+            out.println("<h1>Servlet CreateFeedback at " + request.getContextPath () + "</h1>");
             out.println("</body>");
             out.println("</html>");
         }
@@ -60,30 +56,7 @@ public class ViewListFeedback extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
     throws ServletException, IOException {
-        FeedbackDAO daoFeedback = new FeedbackDAO();
-        String customerName = request.getParameter("customerName");
-        String imageAndVideo = request.getParameter("imageAndVideo");
-        request.setAttribute("customerName", customerName);
-        request.setAttribute("imageAndVideo", imageAndVideo);
-        //======phan trang
-        int totalPages = daoFeedback.getTotalFeedback(customerName,imageAndVideo);
-        int endPage = totalPages/7;
-        if(totalPages % 7 !=0){
-            endPage  ++;
-        }
-        request.setAttribute("endPage", endPage);
-        String indexStr  = request.getParameter("index");
-        int index = 1;
-        try {
-            index = Integer.parseInt(indexStr);
-        } catch (Exception e) {
-            
-        }
-        ArrayList<Feedback> listFeedback = daoFeedback.getAllFeedback(customerName,imageAndVideo,index);
-
-        //======end phan trang
-        request.setAttribute("listFeedback", listFeedback);
-        request.getRequestDispatcher("viewListFeedback.jsp").forward(request, response);
+        response.sendRedirect("CreateFeedback.jsp");
     } 
 
     /** 
@@ -96,29 +69,14 @@ public class ViewListFeedback extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
     throws ServletException, IOException {
-        FeedbackDAO daoFeedback = new FeedbackDAO();
-        
-        String customerName = request.getParameter("customerName");
-        String imageAndVideo = request.getParameter("imageAndVideo");
-        request.setAttribute("customerName", customerName);
-        request.setAttribute("imageAndVideo", imageAndVideo);
-        //phan trang
-        int totalPages = daoFeedback.getTotalFeedback(customerName,imageAndVideo);
-        int endPage = totalPages/7;
-        if(totalPages % 7 !=0){
-            endPage  ++;
-        }
-        request.setAttribute("endPage", endPage);
-        String indexStr  = request.getParameter("index");
-        int index = 1;
-        try {
-            index = Integer.parseInt(indexStr);
-        } catch (Exception e) {
-            
-        }
-        ArrayList<Feedback> listFeedback = daoFeedback.getAllFeedback(customerName,imageAndVideo,index);
-        request.setAttribute("listFeedback", listFeedback);
-        request.getRequestDispatcher("viewListFeedback.jsp").forward(request, response);
+        String note = request.getParameter("note");
+        String customerId = request.getParameter("customerId");
+        String warrantyCardId = request.getParameter("warrantyCardId");
+        FeedbackDAO dao = new FeedbackDAO();
+        dao.createFeedback("2", "7", note);
+        System.out.println(note);
+        // tam thoi ve trang view list de kiem tra sau nay sua
+        response.sendRedirect("ViewListFeedback");
     }
 
     /**
